Refactor tests and updated docs
